Reduce support to PHP 7.4.1 onwards
PHP 7.4.1 broke backward compatibility in which public typed properties no longer trigger `__get`
calls when not initialized, and must be explicitly unset instead. Therefore, we need to drop support
for PHP 7.4.0 in order to write adjustments that are compatible with PHP 7.4.1 onwards, and that
allow lazy-loading of public typed properties.

Ref: https://bugs.php.net/bug.php?id=79373