Refactored #642 to forward the original class to `PublicScopeSimulator` when generating notice emulation code
Before this, the decision whether an interface was being proxied was delegated to the magic method generator,
which led to an explosion in conditionals in upstream code. Here we instead pass the whole `ReflectionClass`
for the parent symbol to the `PublicScopeSimulator`, therefore adding a single conditional to the source of
`PublicScopeSimulator`, leading to reduced complexity.